set account status and skip handling of offline accounts
Initialize account status to "offline" and set it to "online" when the
xmpp connection is ready. Skip handling of messages from a client or
updating the buddy list when the account is in the "offline" state to
avoid issues with sending commands or messages over an offline or
uninitialized xmpp connection.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>